+++ auth-db/queries/email-auth/index.ts
@@ -3,3 +3,4 @@ export * from "./get-by-email.ts";
 export * from "./update-verification.ts";
 export * from "./update-forgot-password-token.ts";
 export * from "./update-password-hash.ts";
+export * from "./get-by-forgot-password-token.ts";
